% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_dbconnection.R
\name{access_dbconnection}
\alias{access_dbconnection}
\title{Connection with Access database (with a JDBC driver)}
\usage{
access_dbconnection(driver_name = "u_can_access", access_db_loc)
}
\arguments{
\item{driver_name}{Access' driver name. By default "u_can_access". You can also choose the driver "access_jdbc42".}

\item{access_db_loc}{Location of the Access database (for example Directory_1\%Directory_2\%Directory_X\%access_DB.mdb).}
}
\value{
The function return a R object with Access database identification of connection.
}
\description{
Generate a common connection with an Access database using a JDBC driver.
}
\references{
\url{https://github.com/OB7-IRD/furdeb}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
