% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdth_position_conversion.R
\name{avdth_position_conversion}
\alias{avdth_position_conversion}
\title{AVDTH database position conversion}
\usage{
avdth_position_conversion(data, latitude, longitude, quadrant)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. A R data frame.}

\item{latitude}{{\link[base]{character}} expected. Column name of latitude data.}

\item{longitude}{{\link[base]{character}} expected. Column name of longitude data.}

\item{quadrant}{{\link[base]{character}} expected. Column name of quadrant data.}
}
\value{
This function add two column to the input data frame, longitude_dec and latitude_dec, with longitude and latitude data in decimal format.
}
\description{
Convert position format in AVDTH to decimal format.
}
