% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cwp_grid.R
\name{create_cwp_grid}
\alias{create_cwp_grid}
\title{Creates a CWP (FAO Coordinating Working Party) grid spatial object}
\usage{
create_cwp_grid(
  resolution = "1deg_x_1deg",
  longitude_min = -180,
  latitude_min = -90,
  longitude_max = 180,
  latitude_max = 90,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{resolution}{{\link[base]{character}} expected. A string matching one of the accepted resolution values. Accepted resolutions values are "10min_x_10min", "20min_x_20min", "30min_x_30min", "30min_x_1deg", "1deg_x_1deg", "5deg_x_5deg", "10deg_x_10deg", "20deg_x_20deg" and "30deg_x_30deg".}

\item{longitude_min}{{\link[base]{integer}} expected. Longitude minimum value of the output grid in decimal degree. By default -180.}

\item{latitude_min}{{\link[base]{integer}} expected. Latitude minimum value of the output grid in decimal degree. By default -90.}

\item{longitude_max}{{\link[base]{integer}} expected. Longitude maximum value of the output grid in decimal degree. By default 180.}

\item{latitude_max}{{\link[base]{integer}} expected. Latitude maximum value of the output grid in decimal degree. By default 90.}

\item{parallel}{{\link[base]{logical}} expected. Run in parallel (if you have at least two processor cores). By default FALSE.}

\item{...}{Others parallel options}
}
\value{
Create an object of class "SpatialPolygonsDataFrame" named grid_cwp_resolution (where resolution is the function argument).
}
\description{
Creates a CWP grid spatial object.
}
\references{
CWP Handbook - http://www.fao.org/fishery/cwp/handbook/G/en
From a function developed by Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}.
}
