% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpw2center.R
\name{cwp2center}
\alias{cwp2center}
\title{Conversion of CWP to latitude and longitude (related to cwp centroid)}
\usage{
cwp2center(data, cwp_name, cwp_length)
}
\arguments{
\item{data}{A R data frame with at least one column with cwp data.}

\item{cwp_name}{Column name of data data in text format.}

\item{cwp_length}{Length of cwp. For example, for a square of 1°x1° enter 1.}
}
\value{
This function add two column to the input data frame, longitude_dec and latitude_dec, with longitude and latitude data in decimal format.
}
\description{
Conversion of CWP to latitude and longitude (in decimal degrees). Be careful, latitude and longitude are related to the cwp centroid.
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Antoine Duparc, \email{antoie.duparc@ird.fr}

Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
