% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpw2latlon.R
\name{cwp2latlon}
\alias{cwp2latlon}
\title{Conversion of CWP to latitude and longitude (related to cwp corner)}
\usage{
cwp2latlon(data, cwp_name)
}
\arguments{
\item{data}{A R data frame with at least one column with cwp data.}

\item{cwp_name}{Column name of data data in text format.}
}
\value{
This function add two column to the input data frame, longitude_dec and latitude_dec, with longitude and latitude data in decimal format.
}
\description{
Conversion of CWP to latitude and longitude (in decimal degrees). Be careful, latitude and longitude are related to the cwp corner nearest the zero (latitude = 0 and longitude = 0).
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Antoine Duparc, \email{antoie.duparc@ird.fr}

Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
