% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwp_to_center.R
\name{cwp_to_center}
\alias{cwp_to_center}
\title{Conversion of cwp to latitude and longitude (related to cwp centroid)}
\usage{
cwp_to_center(data, cwp_name, cwp_length)
}
\arguments{
\item{data}{(data.frame) A R data frame with at least one column with cwp data.}

\item{cwp_name}{(character) Column name of cwp data.}

\item{cwp_length}{(numeric) Length of cwp. For example, for a square of 1°x1° enter 1.}
}
\value{
This function add four columns to the input data frame: cwp, quadrat, longitude_dec and latitude_dec (with longitude and latitude data in decimal format).
}
\description{
Conversion of CWP to latitude and longitude (in decimal degrees). Be careful, latitude and longitude are related to the cwp centroid.
}
