% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwp_to_lat_lon.R
\name{cwp_to_lat_lon}
\alias{cwp_to_lat_lon}
\title{Conversion of cwp to latitude and longitude (related to cwp corner)}
\usage{
cwp_to_lat_lon(data, cwp_name)
}
\arguments{
\item{data}{(data.frame) A R data frame with at least one column with cwp data.}

\item{cwp_name}{(character) Column name of cwp data.}
}
\value{
This function add four columns to the input data frame: cwp, quadrat, longitude_dec and latitude_dec (with longitude and latitude data in decimal format).
}
\description{
Conversion of CWP to latitude and longitude (in decimal degrees). Be careful, latitude and longitude are related to the cwp corner nearest the zero (latitude = 0 and longitude = 0).
}
