% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{Connection to a PostgreSQL database}
\usage{
db_connection(db_user, db_password, db_dbname, db_host, db_port)
}
\arguments{
\item{db_user}{Login or username for the database connection.}

\item{db_password}{Password for the database connection.}

\item{db_dbname}{Name/identification of the database.}

\item{db_host}{Host adress of the database.}

\item{db_port}{Identification of the port.}
}
\value{
The function return a format class PosstgreSQLConnection in the R's global envrionment.
}
\description{
This function able to establish a connection with a PostgreSQL database.
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
