% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_area_overlay.R
\name{fao_area_overlay}
\alias{fao_area_overlay}
\title{Consistent spatial fao area overlay}
\usage{
fao_area_overlay(data, overlay_level = "major", longitude_name,
  latitude_name)
}
\arguments{
\item{data}{R dataframe, with at least two columns with longitude and latitude values. Be careful! Your longitude and latitude data have to be in the WGS84 projection and coordinates in decimal degrees.}

\item{overlay_level}{Maximum fao area level of accuarcy that you want for classified your data (character value). By default, major fao fishing area are selected. Check the section details below.}

\item{longitude_name}{Longitude column name in your data (character value).}

\item{latitude_name}{Latitude column name in your data (character value).}
}
\value{
The function return your input dataframe with one or several columns (regarding specification in the argument "overlay_level") which contains fao area classification.
}
\description{
Consistent spatial fao area overlay for points, grids and polygons.
}
\details{
For the argument "overlay_level", you can choose between 5 modalities (descending size classification):
\itemize{
 \item{major: }{major fao fishing area}
 \item{subarea: }{sub fao fishing area}
 \item{division: }{division fao fishing area}
 \item{subdivision: }{sub-division fao fishing area}
 \item{subunit: }{sub-unit fao fishing area}
}
All the items above your specification (thus contain it at higher levels) will be added in the output. For example, if you select "subarea", you will also have the information about the major area concerning.
If you want more informations visit http://www.fao.org/fishery/area/search/en
}
\examples{
# Example for classification until division fao fishing area
tmp <- fao_area_overlay(data = data,
                        overlay_level = "division",
                        longitude_name = "longitude",
                        latitude_name = "latitude")
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Antoine Duparc, \email{antoine.duparc@ird.fr}

Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
