% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_area_overlay_unassociated.R
\name{fao_area_overlay_unassociated}
\alias{fao_area_overlay_unassociated}
\title{Consistent spatial unassociated fao area overlay}
\usage{
fao_area_overlay_unassociated(data, overlay_level = "major",
  longitude_name, latitude_name, tolerance = 500)
}
\arguments{
\item{data}{R dataframe, with at least two columns with longitude and latitude values. Be careful! Your longitude and latitude data have to be in the WGS84 projection and coordinates in decimal degrees.}

\item{overlay_level}{Maximum fao area level of accuarcy that you want for classified your data (character value). By default, major fao fishing area are selected. Check the section details below.}

\item{longitude_name}{Longitude column name in your data (character value).}

\item{latitude_name}{Latitude column name in your data (character value).}

\item{tolerance}{Tolerance of maximum distance between coordinates and FAO area (in km, numerical value expected). By default, 500 km.}
}
\value{
The function return your input dataframe with one or several columns (regarding specification in the argument "overlay_level") which contains fao area classification.
}
\description{
Consistent spatial fao area overlay for points, grids and polygons. This function work similar than the function "fao_area_overlay" but with a tolerance for coordinates onshore (for the case of coastal fishing and coordinates estimated).
}
\details{
For the argument "overlay_level", you can choose between 5 modalities (descending size classification):
\itemize{
 \item{major: }{major fao fishing area}
 \item{subarea: }{sub fao fishing area}
 \item{division: }{division fao fishing area}
 \item{subdivision: }{sub-division fao fishing area}
 \item{subunit: }{sub-unit fao fishing area}
}
All the items above your specification (thus contain it at higher levels) will be added in the output. For example, if you select "subarea", you will also have the information about the major area concerning.
If you want more informations visit http://www.fao.org/fishery/area/search/en
}
\examples{
# Example for classification until division fao fishing area, with a tolerance of 50 km
tmp <- fao_area_overlay(data = data,
                        overlay_level = "division",
                        longitude_name = "longitude",
                        latitude_name = "latitude",
                        tolerance = 50)
}
\references{
\url{https://github.com/OB7-IRD/furdeb}
}
\author{
Antoine Duparc, \email{antoine.duparc@ird.fr}

Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
