% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_cwp_shape_manipulation.R
\name{fao_cwp_shape_manipulation}
\alias{fao_cwp_shape_manipulation}
\title{Optimisation function for FAO CWP shape file manipulation}
\usage{
fao_cwp_shape_manipulation(referential_grid_file_path, output_directory_path)
}
\arguments{
\item{referential_grid_file_path}{{\link[base]{character}} expected. File path of the FAO CWP shape file (shp extension expected).}

\item{output_directory_path}{{\link[base]{character}} expected. Directory path for the output.}
}
\description{
Optimise FAO CWP shape file and save it into a RData format for increase the reading process.
}
