% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_mode_code_to_name.R
\name{fishing_mode_code_to_name}
\alias{fishing_mode_code_to_name}
\title{Fishing mode(s) name(s) creation}
\usage{
fishing_mode_code_to_name(fishing_mode_code)
}
\arguments{
\item{fishing_mode_code}{Fishing mode code(s) code(s) in numerical values.}
}
\value{
A character vector in relation with the fishing mode(s) code(s) provided.
}
\description{
Fishing mode(s) name(s) creation in relation with fishing mode referential of the IRD Ob7 (Observatory of Exploited Tropical Pelagic Ecosystems).
}
\examples{
fishing_mode_code_to_name(fishing_mode_code = c(1, 2, 3))
}
