% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furdeb-package.R
\docType{package}
\name{furdeb-package}
\alias{furdeb}
\alias{furdeb-package}
\title{furdeb package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions and processes for R development and beyond!
}
\section{Databases manipulations}{

\itemize{
\item{\code{\link{access_dbconnection}}}
\item{\code{\link{postgresql_dbconnection}}}
}
}

\section{Spatial design}{

\itemize{
\item{\code{\link{avdth_position_conversion_dec}}}
\item{\code{\link{fao_reference_grid_update}}}
\item{\code{\link{lat_lon_cwp_manipulation}}}
\item{\code{\link{lat_long_to_csquare}}}
\item{\code{\link{marine_area_overlay}}}
}
}

\section{Data manipulations}{

\itemize{
\item{\code{\link{list_logical_merge}}}
}
}

\section{Setup help}{

\itemize{
\item{\code{\link{configuration_file}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/furdeb}
  \item Report bugs at \url{https://github.com/OB7-IRD/furdeb/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} [copyright holder]

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr}
  \item Antoine Duparc \email{antoine.duparc@ird.fr}
}

Other contributors:
\itemize{
  \item Norbert Billet \email{norbert.billet@ifremer.fr} [contributor]
}

}
\keyword{internal}
