% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furdeb.R
\docType{package}
\name{furdeb}
\alias{furdeb}
\alias{furdeb-package}
\title{furdeb package}
\description{
Functions and processes for R development
}
\section{furdeb functions}{

\itemize{
\item{\code{\link{access_dbconnection}}}
\item{\code{\link{avdth_position_conversion_dec}}}
\item{\code{\link{configuration_file}}}
\item{\code{\link{db_connection}}}
\item{\code{\link{mdb2Sqlite3}}}
\item{\code{\link{sql_inset}}}
\item{\code{\link{fao_area_overlay}}}
\item{\code{\link{fao_area_overlay_unassociated}}}
\item{\code{\link{lat_long_to_csquare}}}
\item{\code{\link{cwp_to_center}}}
\item{\code{\link{cwp_to_lat_lon}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/furdeb}
  \item Report bugs at \url{https://github.com/OB7-IRD/furdeb/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@ird.fr}

}
