% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_corr.R
\name{ggplot_corr}
\alias{ggplot_corr}
\title{Correlogram function}
\usage{
ggplot_corr(
  data,
  lag_max = as.integer(24),
  ci = as.numeric(0.95),
  large_sample_size = FALSE,
  ...
)
}
\arguments{
\item{data}{Time series object, {\link[base]{numeric}} or {\link[base]{matrix}} expected. A univariate or multivariate (not ccf) numeric time series object or a numeric vector or matrix.}

\item{lag_max}{{\link[base]{integer}} expected. Maximum lag at which to calculate the acf. By default 24.}

\item{ci}{{\link[base]{numeric}} expected. Value for the confidence interval. By default 0.95.}

\item{large_sample_size}{{\link[base]{logical}} expected. Specify here if you have a large sample. By default FALSE.}
}
\value{
Return a list with two elements who contains plots, in ggplot format, of correlation and partial correlation.
}
\description{
Correlogram function for both large/small sample size. This function if based on the working development of Kevin Liu. If you want to know, take a look on his website https://rh8liuqy.github.io/ACF_PACF_by_ggplot2.html.
}
