% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwp_to_latitude_longitude.R
\name{latitude_longitude_to_cwp}
\alias{latitude_longitude_to_cwp}
\title{Conversion of latitude and longitude (in decimal degrees) to cwp.}
\usage{
latitude_longitude_to_cwp(data, latitude_name, longtitude_name, cwp_length)
}
\arguments{
\item{data}{(data.frame) A R data frame with at least two column with latitude and longtitude data.}

\item{latitude_name}{(character) Column name of latitude data (display in decimal degrees).}

\item{longtitude_name}{(character) Column name of longitude data (display in decimal degrees).}

\item{cwp_length}{(numeric) Length of cwp. For example, for a square of 1°x1° enter 1.}
}
\description{
Conversion of latitude and longitude (in decimal degrees) to cwp.
}
