% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_logical_merge.R
\name{list_logical_merge}
\alias{list_logical_merge}
\title{Logical merging for list}
\usage{
list_logical_merge(first_list, second_list)
}
\arguments{
\item{first_list}{{\link[base]{list}} expected. First list to merge.}

\item{second_list}{{\link[base]{list}} expected. Second list to merge.}
}
\description{
Process for merging two lists according to the name of them.
}
