% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marine_area_overlay.R
\name{marine_area_overlay}
\alias{marine_area_overlay}
\title{Consistent spatial marine area overlay}
\usage{
marine_area_overlay(
  data,
  overlay_expected,
  longitude_name,
  latitude_name,
  fao_area_file_path = NULL,
  fao_overlay_level = NULL,
  auto_selection_fao = FALSE,
  eez_area_file_path = NULL,
  for_fdi_use = NULL,
  ices_area_file_path = NULL,
  silent = FALSE
)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. R dataframe, with at least two columns with longitude and latitude values. Be careful! Your longitude and latitude data have to be in the WGS84 projection and coordinates in decimal degrees.}

\item{overlay_expected}{{\link[base]{character}} expected. Type of overlay output. You can choose between "fao_area", "eez_area", "fao_eez_area", "ices_area" or "all".}

\item{longitude_name}{{\link[base]{character}} expected. Longitude column name in your data.}

\item{latitude_name}{{\link[base]{character}} expected. Latitude column name in your data.}

\item{fao_area_file_path}{{\link[base]{character}} expected. File path of the fao area shape. You can provide a .shp or a .Rdata/.RData file.}

\item{fao_overlay_level}{{\link[base]{character}} expected. Level of fao accuarcy that you want for classified your data. By default, major fao fishing area are selected. Check the section details below.}

\item{auto_selection_fao}{{\link[base]{logical}} expected. Add a new column in the output with the most detailed overlay level available.}

\item{eez_area_file_path}{{\link[base]{character}} expected. File path of the eez area shape. You can provide a .shp or a .Rdata/.RData file.}

\item{for_fdi_use}{{\link[base]{logical}} expected. Add a new column in the output with the FDI variable "eez_indicator".}

\item{ices_area_file_path}{{\link[base]{character}} expected. File path of the ices area shape. You can provide a .shp or a .Rdata/.RData file.}

\item{silent}{{\link[base]{logical}} expected. Display or not warning information regarding projection of your spatial coordinates. By default FALSE.}
}
\value{
The function return your input data frame with one or several columns (regarding arguments' specification) which contains area classification.
}
\description{
Consistent spatial marine area overlay for points, grids and polygons (related to fao, eez or ices areas).
}
\details{
For the argument "fao_overlay_level", you can choose between 5 modalities (descending size classification):
\itemize{
 \item{ocean: }{ocean area}
 \item{major: }{major fao fishing area}
 \item{subarea: }{sub fao fishing area}
 \item{division: }{division fao fishing area}
 \item{subdivision: }{sub-division fao fishing area}
 \item{subunit: }{sub-unit fao fishing area}
}
Specificity for fao fishing area parameters: all the items above your specification (thus contain it at higher levels) will be added in the output. For example, if you select "subarea", you will also have the information about the major area concerning.
If you want more information visit http://www.fao.org/fishery/area/search/en
}
