% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb2Sqlite3.R
\name{mdb2Sqlite3}
\alias{mdb2Sqlite3}
\title{Converting an Access database to an SQLite3 database}
\usage{
mdb2Sqlite3(access_con, target_dir, target_file_name, verbose = TRUE)
}
\arguments{
\item{access_con}{Object with Access database connection identification.}

\item{target_dir}{Directory where the SQLite3 file will be created.}

\item{target_file_name}{SQLite3 file name (for example: database_name.sqlite3).}

\item{verbose}{Should the script be verbose.}
}
\value{
The function return an SQLite3 database in the location directory. The file path of the SQLite3 file is also available.
}
\description{
Convert Access database to sqlite3 database.
}
\section{Specification}{

Should work on windows and on linux (need testing for this OS). For linux the "mdb-tools" must be installed.
}

\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Norbert Billet, \email{norbert.billet@ifremer.fr}

Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
