% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_code_to_name.R
\name{ocean_code_to_name}
\alias{ocean_code_to_name}
\title{Ocean(s) name(s) creation}
\usage{
ocean_code_to_name(ocean_code)
}
\arguments{
\item{ocean_code}{(numeric) Ocean(s) code(s).}
}
\value{
A list with ocean name(s), ocean character code(s), maximum and minimum latitude and longitude of ocean(s) selected.
}
\description{
Ocean(s) name(s) creation in relation with oceans referential of the IRD Ob7 (Observatory of Exploited Tropical Pelagic Ecosystems).
}
\examples{
ocean_code_to_name(ocean_code = c(1, 2, 3))
}
