% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql_dbconnection.R
\name{postgresql_dbconnection}
\alias{postgresql_dbconnection}
\title{Connection to a PostgreSQL database}
\usage{
postgresql_dbconnection(db_user, db_password, db_dbname, db_host, db_port)
}
\arguments{
\item{db_user}{{\link[base]{character}} expected. Login or username for the database connection.}

\item{db_password}{{\link[base]{character}} expected. Password for the database connection.}

\item{db_dbname}{{\link[base]{character}} expected. Name/identification of the database.}

\item{db_host}{{\link[base]{character}} expected. Host adress of the database.}

\item{db_port}{{\link[base]{numeric}} expected. Identification of the port.}
}
\value{
The function return a list.
}
\description{
This function able to establish a connection with a PostgreSQL database.
}
