% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_inset.R
\name{sql_inset}
\alias{sql_inset}
\title{Interpolate values into a SQL string query}
\usage{
sql_inset(db_type, replacement, pattern, query)
}
\arguments{
\item{db_type}{(character) Type of database related to the query. For now, the function accepts only "access" or "postgresql" databases.}

\item{replacement}{(character) Replacement argument for matched pattern (R vector). If you want to specify multiple arguments, use the function c().}

\item{pattern}{(character) String expression to be matched.}

\item{query}{(character) Vector where matches are sought, or an object which can be coerced by as.character to a character vector.}
}
\value{
A vector contain the query with replacement value(s).
}
\description{
Interpolate values into a SQL string query.
}
\examples{
For a query on an Access database
\dontrun{
#' final_query <- sql_inset(db_type = "access",
                            replacement = 2017,
                            pattern = "year_interpolate",
                            query = "intial_query")}
}
