% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_dbconnection.R
\name{access_dbconnection}
\alias{access_dbconnection}
\title{Connection with Access database}
\usage{
access_dbconnection(
  driver_name = "u_can_access",
  access_database_file_path,
  access_jdbc42_driver_file_path = NULL
)
}
\arguments{
\item{driver_name}{{\link[base]{character}} expected. By default "u_can_access" (free driver). Access driver name. You can also choose the driver "access_jdbc42" (paid driver).}

\item{access_database_file_path}{{\link[base]{character}} expected. file path of the Access database (.mdb or .accdb expected).}

\item{access_jdbc42_driver_file_path}{{\link[base]{character}} expected. By default NULL. File path of the access_jdbc42 driver (.jar file expected). Mandatory if "access_jdbc42" is selected in the argument "driver_name".}
}
\value{
The function return a R object with Access database identification of connection.
}
\description{
Generate a common connection with an Access database using a JDBC driver.
}
\details{
Difference between drivers "u_can_access" and "access_jdbc42":
\itemize{
\item{"u_can_access": the main advantage is it's a free java JDBC driver. To understand briefly the process behind, the Access database is converted in HSQLDB system. This conversion could take a long time, especially if the database if large. Furthermore, after the conversion, queries should be faster than if there run on an Access database.}
\item{"access_jdbc42": this driver is paying. In opposition with the "u_can_access" driver, the Access database is not converted. The connection should be faster but the queries could be longer.}
}
}
