% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_dbconnection.R
\name{access_dbconnection}
\alias{access_dbconnection}
\title{Connection with Access database (with a JDBC driver)}
\usage{
access_dbconnection(jdbc_access_driverclass, jdbc_access_driver_loc,
  access_db_loc)
}
\arguments{
\item{jdbc_access_driverclass}{Identification of the Access driver class (for example com.hxtt.sql.access.AccessDriver).}

\item{jdbc_access_driver_loc}{Location of the Access JDBC driver (for example Directory_1\%Directory_2\%Directory_X\%access_JDBCdriver.jar).}

\item{access_db_loc}{Location of the Access database (for example Directory_1\%Directory_2\%Directory_X\%access_DB.mdb).}
}
\value{
The function return a R object with Access database identification of connection.
}
\description{
Generate a common connection with an Access database using a JDBC driver.
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
