% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdth_position_conversion_dec.R
\name{avdth_position_conversion_dec}
\alias{avdth_position_conversion_dec}
\title{AVDTH position conversion to decimal value}
\usage{
avdth_position_conversion_dec(data, latitude, longitude, quadrant)
}
\arguments{
\item{data}{A R data frame.}

\item{latitude}{Column name of latitude data in text format.}

\item{longitude}{Column name of longitude data in text format.}

\item{quadrant}{Column name of quadrant data in text format.}
}
\value{
This function add two column to the input data frame, longitude_dec and latitude_dec, with longitude and latitude data in decimal format.
}
\description{
Convert position format in AVDTH to decimal format.
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
