% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_file.R
\name{configuration_file}
\alias{configuration_file}
\title{Configuration file}
\usage{
configuration_file(new_config = T, path_config)
}
\arguments{
\item{new_config}{(logical) Provide FALSE/F value is you want to use an existing configuration file. By default, the function create a new configuration file (TRUE/T argument value).}

\item{path_config}{(character) Path of the configuration file that you want to use.}
}
\value{
The function returns a list of configuration objects. If you have create a new configuration file, you can export your own configuration file (.csv format with separator ";").
}
\description{
Apply several options to R from a configuration file (able to create one if necessary).
}
\examples{
# If you want to create a new configuration file
\dontrun{
configuration_file()}
# If you want to use an existing configuration file
\dontrun{
configuration_file(new_config = F,
                   path_config = "path_of_your_own_configuration_file")}
}
