% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_file.R
\name{configuration_file}
\alias{configuration_file}
\title{Configuration file}
\usage{
configuration_file(new_configtype = TRUE, path_configtype)
}
\arguments{
\item{new_configtype}{A logical vector: TRUE, FALSE, T or F. Provide false value is you want to use an existing configuration file. By default, the function create a new configuration file (=TRUE).}

\item{path_configtype}{Path (character) of the configuration file that you want to use.}
}
\value{
The function returns a list of configuration objects. If you have create a new configuration file, you can export your own configuration file (.csv format with separator ";").
}
\description{
Apply several options to R from a configuration file (able to create one if necessary).
}
\examples{
# If you want to create a new configuration file
configuration_file()
# If you want to use an existing configuration file
configuration_file(new_configtype = FALSE,
                   path_configtype = "path_of_your_own_configuration_file")
}
\references{
\url{https://github.com/OB7-IRD/furdeb}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
