% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_file.R
\name{configuration_file}
\alias{configuration_file}
\title{Configuration file}
\usage{
configuration_file(configtype)
}
\arguments{
\item{configtype}{Vector of character. For details see below the specification section.}
}
\value{
The function returns a list of configuration objects. If none configuration file has been provided at the beginning, you can export your own configuration file (.csv format).
}
\description{
Apply several options to R from a configuration file (able to create one if necessary).
}
\section{Specification}{

You can specify two kinds of configurations:
\itemize{
\item{"base" for a base configuration without a database's connection}
\item{"db" for an advanced configuration with a database's connection}
}
}

\examples{
#If you want to use basic configuration without a database's connection
configuration_file("base")

#If you want to use advanced configuration with a database's connection
configuration_file("db")
}
\references{
\url{https://github.com/OB7-IRD/toolbox}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
