% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{data_extraction}
\alias{data_extraction}
\title{Data extraction}
\usage{
data_extraction(
  type,
  file_path,
  database_connection = NULL,
  anchor = NULL,
  column_name = NULL,
  column_type = NULL,
  export_path = NULL
)
}
\arguments{
\item{type}{{\link[base]{character}} expected. dabase or csv.}

\item{data_connection}{{\link[base]{list}} expected. Output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the data_extraction function.}

\item{sql_name}{{\link[base]{character}}. Name of the sql file, in fishi.}

\item{csv_path}{{\link[base]{character}}. Path of the csv.}

\item{time_period}{{\link[base]{integer}} expected. Period identification in year.}

\item{country}{{\link[base]{integer}} expected. Country codes identification. [\href{https://ob7-ird.github.io/fishi/articles/Referentials.html}{Country referentials}]}

\item{vessel_type}{{\link[base]{integer}} expected. Vessel type codes identification. [\href{https://ob7-ird.github.io/fishi/articles/Referentials.html}{Vessel_type referentials}]}

\item{vessel_type_select}{{\link[base]{character}} expected. engin or vessel_type.}

\item{ocean}{{\link[base]{integer}} expected. Ocean codes identification. [\href{https://ob7-ird.github.io/fishi/articles/Referentials.html}{Oscean referentials}]}

\item{csv_output}{TRUE to obtain a csv document. FALSE by default.}
}
\value{
The function return ggplot R plot.
}
\description{
Extracts a dataset from a database or csv
}
