% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{Connection to a PostgreSQL database}
\usage{
db_connection(db_user, db_password, db_dbname, db_host, db_port)
}
\arguments{
\item{db_user}{(character) Login or username for the database connection.}

\item{db_password}{(character) Password for the database connection.}

\item{db_dbname}{(character) Name/identification of the database.}

\item{db_host}{(character) Host adress of the database.}

\item{db_port}{(numeric) Identification of the port.}
}
\value{
The function return a format class PosstgreSQLConnection in the R's global envrionment.
}
\description{
This function able to establish a connection with a PostgreSQL database.
}
