% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furdeb-package.R
\docType{package}
\name{furdeb-package}
\alias{furdeb}
\alias{furdeb-package}
\title{FUnctions for R DEvelopment and Beyond}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions and processes for R development and beyond!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/furdeb}
  \item Report bugs at \url{https://github.com/OB7-IRD/furdeb/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} [copyright holder]

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr}
  \item Antoine Duparc \email{antoine.duparc@ird.fr}
}

Other contributors:
\itemize{
  \item Norbert Billet \email{norbert.billet@ifremer.fr} [contributor]
}

}
\keyword{internal}
