% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furdeb-package.R
\docType{package}
\name{furdeb-package}
\alias{furdeb}
\alias{furdeb-package}
\title{furdeb: FUnctions for R DEvelopment and Beyond!}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions and processes for R development and beyond!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/furdeb}
  \item Report bugs at \url{https://github.com/OB7-IRD/furdeb/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Authors:
\itemize{
  \item Oceane Bouhineau \email{oceane.bouhineau@ird.fr} (\href{https://orcid.org/0000-0001-7418-2925}{ORCID})
  \item Clara Lerebourg \email{clara.lerebourg@ird.fr} (\href{https://orcid.org/0000-0003-0589-8932}{ORCID})
}

Other contributors:
\itemize{
  \item Norbert Billet \email{norbert.billet@ifremer.fr} [contributor]
  \item Philippe Sabarros \email{philippe.sabarros@ird.fr} (\href{https://orcid.org/0000-0001-6135-9015}{ORCID}) [contributor]
  \item IRD - Ob7 Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} (\href{https://orcid.org/0000-0002-4642-3418}{ORCID}) [copyright holder, funder]
}

}
\keyword{internal}
