% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furdeb.R
\docType{package}
\name{furdeb}
\alias{furdeb}
\alias{furdeb-package}
\title{furdeb package}
\description{
Functions and processes for R development and beyond!
}
\section{Diverse}{

\itemize{
\item{\code{\link{configuration_file}}}
}
}

\section{Databases manipulations and connections}{

\itemize{
\item{\code{\link{access_dbconnection}}}
\item{\code{\link{db_connection}}}
\item{\code{\link{mdb2Sqlite3}}}
\item{\code{\link{sql_inset}}}
}
}

\section{Spatial manipulations and design}{

\itemize{
\item{\code{\link{avdth_position_conversion_dec}}}
\item{\code{\link{fao_area_overlay}}}
\item{\code{\link{fao_area_overlay_unassociated}}}
\item{\code{\link{lat_long_to_csquare}}}
\item{\code{\link{cwp_to_center}}}
\item{\code{\link{cwp_to_lat_lon}}}
}
}

\section{Referentials}{

\itemize{
\item{\code{\link{fishing_mode_code_to_name}}}
\item{\code{\link{ocean_code_to_name}}}
\item{\code{\link{gear_code_to_name}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/furdeb}
  \item Report bugs at \url{https://github.com/OB7-IRD/furdeb/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@ird.fr}

}
