% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_code_to_name.R
\name{gear_code_to_name}
\alias{gear_code_to_name}
\title{Gear(s) name(s) creation}
\usage{
gear_code_to_name(gear_code)
}
\arguments{
\item{gear_code}{{\link[base]{numeric}} expected. Gear(s) code(s).}
}
\value{
A character vector in relation with the gear(s) code(s) provided.
}
\description{
Gear(s) name(s) creation in relation with gears referential of the IRD Ob7 (Observatory of Exploited Tropical Pelagic Ecosystems).
}
\examples{
gear_code_to_name(gear_code = c(1, 2, 3))
}
