% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_code_to_name.R
\name{ocean_code_to_name}
\alias{ocean_code_to_name}
\title{Ocean(s) name(s) creation}
\usage{
ocean_code_to_name(ocean_code)
}
\arguments{
\item{ocean_code}{Ocean(s) code(s) in numerical values.}
}
\value{
A character vector in relation with the ocean(s) code(s) provided.
}
\description{
Ocean(s) name(s) creation.
}
\references{
\url{https://github.com/OB7-IRD/furdeb}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
