% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pigz.R
\name{save_pigz}
\alias{save_pigz}
\title{Save R Object using pigz compression}
\usage{
save_pigz(
  ...,
  list = character(),
  output_file_path,
  cores_utilisation = "auto",
  compression_level = 1L,
  precheck = TRUE,
  envir = parent.frame(),
  eval_promises = TRUE
)
}
\arguments{
\item{...}{Names of the objects to be saved (as symbols or character strings).}

\item{list}{{\link[base]{character}} expected. A vector containing the names of objects to be saved.}

\item{output_file_path}{{\link[base]{character}} expected. Path name of the file where data will be saved.}

\item{cores_utilisation}{{\link[base]{character}} or {\link[base]{numeric}} expected. Percentage of cores to use to compress (value inferior or equal to 1 expected). Use "auto" for automatic management (number of cores minus 1).}

\item{compression_level}{{\link[base]{integer}} expected. Compression level, 0 for no compression to 9 (maximum).}

\item{precheck}{{\link[base]{logical}} expected. Should the existence of the objects be checked before starting to.}

\item{envir}{Environment to search for objects to be saved.}

\item{eval_promises}{{\link[base]{logical}} expected. Should objects which are promises be forced before saving?}
}
\description{
Save writes an external representation of R objects to the specified file using pigz parallel implementation compression (http://www.zlib.net/pigz/).
}
\details{
Under windows OS utilisation, download through this link (https://sourceforge.net/projects/pigz-for-windows/) the pigz executable file and place it in the System32 directory. The executable file is also available in the package directory (use the function system.file("pigz.zip", package = "furdeb") for located it).
}
