% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_inset.R
\name{sql_inset}
\alias{sql_inset}
\title{Interpolate values into a SQL string query}
\usage{
sql_inset(db_type, replacement, pattern, query)
}
\arguments{
\item{db_type}{Type of database related to the query. For now, the function accepts only "access" or "postgresql" databases.}

\item{replacement}{Replacement argument for matched pattern (R vector). If you want to specify multiple arguments, use the function c().}

\item{pattern}{Character string containing a string expression to be matched.}

\item{query}{A character vector where matches are sought, or an object which can be coerced by as.character to a character vector.}
}
\value{
A vector contain the query with replacement value(s).
}
\description{
Interpolate values into a SQL string query.
}
\examples{
For a query on an Access database
\dontrun{
#' final_query <- sql_inset(db_type = "access",
                         replacement = 2017,
                         pattern = "year_interpolate",
                         query = intial_query)}
}
